requirements.txt
joblib==1.5.3
matplotlib==3.10.8
numpy==2.4.4
pandas==3.0.2
PyYAML==6.0.3
scikit_learn==1.8.0
seaborn==0.13.2
streamlit==1.55.0
